Lisätty mesh luokka

+++ src/vbo.cpp
@@ -7,7 +7,7 @@ Vbo::Vbo(const std::vector<Vertex>& vertices)
 {
     glGenBuffers(1, &ID);
     glBindBuffer(GL_ARRAY_BUFFER, ID);
-    glBufferData(GL_ARRAY_BUFFER, static_cast<long long>(vertices.size() * sizeof(Vertex)), vertices.data(), GL_STATIC_DRAW);
+    glBufferData(GL_ARRAY_BUFFER, (vertices.size() * sizeof(Vertex)), vertices.data(), GL_STATIC_DRAW);
 }
 Vbo::~Vbo()
 {
